Move the SocialMedia.jsx to a separate 'socialMedia' folder.
