+++ src/components/form/ui/Form.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import LoadUnload from "../components/LoadUnload";
 import PersonalInformation from "../components/PersonalInformation";
-import SocialMedia from "../components/SocialMedia";
+import SocialMedia from "../components/socialMedia/ui/SocialMedia";
 import Summary from "../components/Summary";
 import Educations from "../components/education/ui/Educations";
 import WorkExperiences from "../components/workExperience/ui/WorkExperiences";
